Task 2 - Created Order Class
Created an Order class that has the order id, product, and quantity properties. Added a method that refers back to the updatestock method in the Product class to plug in the new this.quantity from the order into the quantity from the product. Added another method to return the details of the order when called.