Task 4 - Implemented Order Management
Create a new method placeOrder to place the new. This is done by finding and ensuring that the product does exist through its ID and then uses an if else statement to ensure that the order can be placed by seeing if there is enough quantity to cover what's requested. This also calculates the total price of the order and then adds the order information to the order array  and logs the order details to the console.