Task 5 - Implemented Product Restocking
Add a method that will restock the stock level for a product by searching for the product in the products array with the product id. Then, if the product exists, the stock for the product will increase by the quantity plugged in when the method was called. Once this is done the console log will return a message saying it has been updated. If the product is not found, there will be a message alerting this error. The updated details for the product will be logged to the console.